+++ apps/web/src/main/java/uk/co/ticklethepanda/carto/apps/web/heatmap/HeatmapParams.java
@@ -7,14 +7,14 @@ import java.time.LocalDateTime;
 
 public class HeatmapParams {
     private final HeatmapDescriptor<LocalDateTime> heatmapDescriptor;
-    private final int pixelSize;
+    private final float pixelSize;
 
-    public HeatmapParams(HeatmapDescriptor<LocalDateTime> heatmapDescriptor, int pixelSize) {
+    public HeatmapParams(HeatmapDescriptor<LocalDateTime> heatmapDescriptor, float pixelSize) {
         this.heatmapDescriptor = heatmapDescriptor;
         this.pixelSize = pixelSize;
     }
 
-    public int getPixelSize() {
+    public float getPixelSize() {
         return pixelSize;
     }
